new lastly

+++ sc-provider/sc-xzsd-pc/src/main/java/com/xzsd/pc/user/services/UserServices.java
@@ -1,17 +1,14 @@
 package com.xzsd.pc.user.services;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
-import com.xzsd.pc.util.PasswordUtils;
-import com.xzsd.pc.util.RandomCode;
+import com.neusoft.core.restful.AppResponse;
+import com.xzsd.pc.user.entity.AdminData;
+import com.xzsd.pc.util.*;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import com.xzsd.pc.user.dao.UserDao;
 import com.xzsd.pc.user.entity.User;
-import com.xzsd.pc.util.ResponceData;
-import com.xzsd.pc.util.ResponceDataState;
 import javax.annotation.Resource;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 /**
@@ -23,7 +20,6 @@ import java.util.List;
 public class UserServices {
     @Resource
     private UserDao userDao;
-    private ResponceData responceData;
     /**
      * 查询用户
      * @param
@@ -31,12 +27,12 @@ public class UserServices {
      * @Author zhc
      * @Date 2020-03-24
      */
-    public ResponceData queryUser(User user){
+    public AppResponse queryUser(User user){
         User appResponse = userDao.queryUserData(user);
         if(null == appResponse){
-            return new ResponceData(ResponceDataState.values()[3].getCode(),"查询为空",null);
+            return AppResponse.bizError("查询为空");
         }
-        return  new ResponceData(ResponceDataState.values()[0].getCode(),"查询成功",appResponse);
+        return AppResponse.success("查询成功!",appResponse);
     }
     /**
      * 增加用户
@@ -44,29 +40,30 @@ public class UserServices {
      * 2020年3月24日19:05:18
      */
     @Transactional(rollbackFor = Exception.class)
-    public ResponceData addUser(User user){
+    public AppResponse addUser(User user){
             //判断参数
             if(null == user.getUserEmail()){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户邮箱参数为空!",null);
+                user.setUserEmail("");
+//                AppResponse.success,"用户邮箱参数为空!",null);
             }
             if(null == user.getUserPass()){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户密码参数为空!",null);
+                return AppResponse.paramError("用户密码参数为空!");
             }
             if(null == user.getUserName()){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户名称为空!",null);
+                return AppResponse.paramError("用户名称为空!");
             }
             if(null == user.getUserAccount()){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户账号为空!",null);
+                return AppResponse.paramError("用户账号为空!");
             }
             if(null == user.getUserPhone()){
-               return new ResponceData(ResponceDataState.values()[3].getCode(),"用户手机号为空!",null);
+               return AppResponse.paramError("用户手机号为空!");
             }
             //先查询 是否存在账号、手机
             if(userDao.countUserAcct(user)>0){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户账号已存在!",null);
+                return AppResponse.paramError("用户账号已存在!");
             }
             if(userDao.countUserPhone(user)>0){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"用户手机号已存在!",null);
+                return AppResponse.paramError("用户手机号已存在!");
             }
             //设置默认头像
             if(null == user.getDefaultImageUrl() || user.getDefaultImageUrl() ==""){
@@ -77,9 +74,9 @@ public class UserServices {
             int result = userDao.addUser(user);
             //如果新增成功，返回结果
             if(result == 0){
-                return new ResponceData(ResponceDataState.values()[3].getCode(),"增加失败",null);
+                AppResponse.bizError("增加失败");
             }
-            return new ResponceData(ResponceDataState.values()[0].getCode(),"增加成功",null);
+            return AppResponse.success("增加成功");
     }
 
     /**
@@ -88,15 +85,27 @@ public class UserServices {
      * 2020年3月24日22:52:47
      */
     @Transactional(rollbackFor = Exception.class)
-    public ResponceData updateUser(User user){
+    public AppResponse updateUser(User user){
         if(null == user.getUserCode()){
-            return new ResponceData(ResponceDataState.values()[3].getCode(),"用户编号为空!",null);
+            return AppResponse.paramError("用户编号为空!");
+        }
+        //更新用户
+        if(userDao.countUserAcct(user) > 0){
+            return AppResponse.paramError("用户账号已存在!");
+        }
+        //设置默认头像
+        if(null == user.getDefaultImageUrl() || user.getDefaultImageUrl() ==""){
+            user.setDefaultImageUrl(RandomCode.defaultImageUrl());
+        }
+        //加密密码
+        if(null != user.getUserPass() && user.getUserPass() != ""){
+            user.setUserPass(PasswordUtils.generatePassword(user.getUserPass()));
         }
         int result = userDao.updateUser(user);
         if(0 == result){
-            return new ResponceData(ResponceDataState.values()[0].getCode(),"更新失败",null);
+            return  AppResponse.bizError("更新失败");
         }
-        return new ResponceData(ResponceDataState.values()[0].getCode(),"更新成功",null);
+        return AppResponse.success("更新成功");
     }
     /**
      * 删除用户
@@ -104,32 +113,50 @@ public class UserServices {
      * 2020年3月25日09:16:14
      */
     @Transactional(rollbackFor = Exception.class)
-     public ResponceData deleteUser(User user){
+     public AppResponse deleteUser(User user){
          //如果有逗号，那么是分割开来
+        if(null == user.getUserCode()){
+           return AppResponse.paramError("用户编号为空!");
+        }
          List<String> listCode = Arrays.asList(user.getUserCode().split(","));
+         if(listCode.size()  == 0){
+            return AppResponse.paramError("删除的数量为空");
+         }
          int result =  userDao.deleteUser(listCode,user.getUserCode());
          if(result > 0){
-             responceData = new ResponceData(ResponceDataState.values()[0].getCode(),"删除成功",null);
+             return AppResponse.success("删除成功!");
          }
-         return new ResponceData(ResponceDataState.values()[3].getCode(),"删除失败",null);
+         return AppResponse.bizError("删除失败");
      }
     /**
      * 用户列表查询
      * author: zhc
      * 2020年3月25日10:56:26
      */
-    public ResponceData queryUserList(User user){
-        //逻辑判断
-        if(user.getPageNum() == 0 || user.getPageSize() == 0){
-            return  new ResponceData(ResponceDataState.values()[3].getCode(),"页号或者页数量参数不能为空",null);
-        }
+    public AppResponse queryUserList(User user){
         //分页
-        PageHelper.startPage(user.getPageSize(),user.getPageNum());
-        List<User> userInfoList = userDao.listUsersByPage(user);
+        PageHelper.startPage(user.getPageNum(),user.getPageSize());
+        List<User> userInfoList = userDao.listUsers(user);
         PageInfo<User> pageData = new PageInfo<User>(userInfoList);
         if(userInfoList.size() == 0){
-            responceData = new ResponceData(ResponceDataState.values()[3].getCode(),"查询为空",userInfoList);
+            return AppResponse.bizError("查询为空");
         }
-        return new ResponceData(ResponceDataState.values()[0].getCode(),"查询成功",pageData);
+        return AppResponse.success("查询成功",pageData);
+    }
+
+    /**
+     * 获取当前登入用户信息
+     * @return
+     */
+    public AppResponse getuserdata(){
+        String userCode = SecurityUtils.getCurrentUserUsername();
+        if(userCode.length() == 0){
+            return AppResponse.bizError("当前没有登入用户!");
+        }
+        AdminData adminData = userDao.getuserdata(userCode);
+        if(null != adminData){
+            return AppResponse.success("获取成功!",adminData);
+        }
+        return AppResponse.bizError("获取失败，请重新登入!");
     }
 }
